python/requirements.txt
flask
flask_cors
pytesseract
Pillow
mysql-connector-python
